refactor: organize mixins & custom selectors
