+++ package.json
@@ -25,6 +25,7 @@
     "zod": "^3.21.4"
   },
   "devDependencies": {
+    "@csstools/postcss-global-data": "^1.0.3",
     "@svelte-put/clickoutside": "^2.0.0",
     "@svelte-put/modal": "^1.0.0",
     "@svelte-put/preprocess-inline-svg": "^1.2.1",
@@ -51,6 +52,7 @@
     "lint-staged": "^13.2.0",
     "npm-run-all": "^4.1.5",
     "postcss": "^8.4.21",
+    "postcss-custom-selectors": "^7.1.2",
     "postcss-html": "^1.5.0",
     "postcss-import": "^15.1.0",
     "postcss-js": "^4.0.1",
