+++ src/lib/client/styles/components/link.css
@@ -41,8 +41,7 @@
     color: theme('colors.svelte');
   }
 
-  &[data-current]:not([data-current="false"]),
-  &[aria-current]:not([aria-current="false"]) {
+  &:--current {
     color: theme('colors.svelte');
   }
 }
