+++ src/lib/client/styles/globals/mixins/index.cjs
@@ -1,0 +1,4 @@
+module.exports = {
+  ...require('./dark.cjs'),
+  ...require('./space.cjs'),
+};
