--- src/lib/client/styles/mixins/space.css
@@ -1,11 +1,0 @@
-/* @define-mixin space-x $value {
-  & > * + * {
-    margin-left: $value;
-  }
-}
-
-@define-mixin space-y $value {
-  & > * + * {
-    margin-top: $value;
-  }
-} */
